requirements/prod.txt
Django
psycopg2
djangorestframework
django-mptt